requirements.txt
matplotlib
torch
torchvision
numpy
scikit-image
argparse
wandb
Pillow
tqdm